third_party/blink/web_tests/editing/inserting/replace-in-heading-001-expected.txt

PASS $("destination").textContent is "source"
PASS successfullyParsed is true

TEST COMPLETE
